Changed ConfigIdInputForm
Redid the input form using tools built in to Visual Studio.
The form now allows for proper checks if input is too large or invalid.

Removed check for correct length at start of sendSC, along with
associated variables.
