+++ CitirocUI/CitirocUI.csproj
@@ -99,8 +99,14 @@
     <Compile Include="Citiroc.Designer.cs">
       <DependentUpon>Citiroc.cs</DependentUpon>
     </Compile>
-    <Compile Include="InputForm.cs">
+    <Compile Include="ConfigInputForm.cs">
       <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="ConfigIdInputForm.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="ConfigIdInputForm.Designer.cs">
+      <DependentUpon>ConfigIdInputForm.cs</DependentUpon>
     </Compile>
     <Compile Include="ProtoCubesNumBinsForm.cs">
       <SubType>Form</SubType>
@@ -173,6 +179,12 @@
     <Compile Include="USB.cs" />
     <EmbeddedResource Include="Citiroc.resx">
       <DependentUpon>Citiroc.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="ConfigInputForm.resx">
+      <DependentUpon>ConfigInputForm.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="ConfigIdInputForm.resx">
+      <DependentUpon>ConfigIdInputForm.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="ProtoCubesMonitor.resx">
       <DependentUpon>ProtoCubesMonitor.cs</DependentUpon>
